Crab/requirements.txt
# Core PyTorch packages - WITHOUT +cu121 suffix (will be handled by index-url)
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0

# Transformers and ML packages
transformers==4.47.1
huggingface-hub==0.27.0
tokenizers==0.21.0
safetensors==0.4.5

# Audio processing
librosa==0.10.1
audioread==3.0.0
soundfile==0.12.1
soxr==0.3.6
praat-parselmouth

# Scientific computing
numpy==1.24.4
scipy==1.11.2
pandas==2.1.0
scikit-learn==1.3.0
numba==0.57.1
llvmlite==0.40.1

# Utilities
joblib==1.3.2
tqdm==4.66.1
filelock==3.12.4
fsspec==2023.9.1
regex==2023.8.8
packaging==23.1
platformdirs==3.10.0

# Data handling
python-dateutil==2.8.2
pytz==2023.3.post1
tzdata==2023.3
six==1.16.0

# Visualization
matplotlib==3.9.0
Pillow==9.4.0

# Serialization
PyYAML==6.0.1
msgpack==1.0.5

# Compression
brotlipy==0.7.0

# Other dependencies
decorator==5.1.1
lazy_loader==0.3
pooch==1.7.0
threadpoolctl==3.2.0
umap-learn==0.5.7

# Development tools
jupyter